+++ src/test/kotlin/lotto/domain/NumberIssuerTest.kt
@@ -1,6 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.NumberIssuer
 import lotto.utils.Constant
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
@@ -24,8 +23,6 @@ class NumberIssuerTest {
 
     @Test
     fun `구매한 로또의 갯수만큼 로또를 발행해 줘야 한다`() {
-        Assertions.assertThat(FIVE_LOTTOS).isEqualTo(NumberIssuer.issueNumbers(5000))
+        Assertions.assertThat(FIVE_LOTTOS).isEqualTo(NumberIssuer.issueNumbers(5000).size)
     }
-
-
 }
