+++ src/test/kotlin/lotto/domain/WinningLottoTest.kt
@@ -1,9 +1,5 @@
-package lotto
+package lotto.domain
 
-import lotto.domain.Lotto
-import lotto.domain.LottoNumber
-import lotto.domain.WinningLotto
-import lotto.domain.WinningRank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
